requirements.txt
paho-mqtt>=1.6.1
pyaudio>=0.2.13
opuslib>=3.0.1
pynput>=1.7.6
requests>=2.31.0
pyyaml>=6.0.1
cryptography>=41.0.0 